job:#10345 Account for pyrsl change to handle blanks
Work done to move BP to pyrsl 2.1.0 (#10338 and #8442) made a change to
how referential attributes were handled. This had some fall out whereby
cme.Specialism (previously sometimes empty) was filled with "--". These
places had to be updated to watch for Specialism="==" and handle it.
This is a place that was missed when that work was done. It simply
affects the test generation.