job:#10345 Changes to account for differences in git project import
The import of projects from zip files has been broken since the move to
Eclipse Oxygen. The BridgePoint test convention for git tests is to
zip-up a git repository that contains model(s). Under the zip file and
folders that represent git repositories (they hold a ".git" file and
under this folder is then a single model file. The convention used is
that the git repository name matches the eclipse project name. I updated
this class to account for this convention by assuring that during import
the model folder is selected for import and NOT the git repository. To
do this I used the git view and looked under the "Working Folder" to
find  (and select) the model folder. This change fixed the tests to
account for a git import behavior change that was introduced by the
updated git plugin in the Oxygen release.