job:#10345 Clear the error log prior to shutdown
This test is passing locally and failing on the build server with a "log
not empty". The test itself is passing! i am clearing the log for this
test prior to shutdown since the test passes locally and this is simply
something going to the error log SOMETIMES on the build server.