job:#10345 Enhanced the DeadlockDetector
I modified the deadlock detector to not just catch deadlocks but to also
handle the time limit exceeded error. This handler checks for deadlock
first, and reports that if it is seen, then if checks timelimit
exceeded. In the time-limit exceeded situation we dump ALL threads
(because we don't know what happened).