job #12112 - Added default key letter rename tests
Two tests added to RenameTest2Generics.java; the first checks to see if
a defaulted key letter value is changed on element rename, and the
second checks to ensure that an undefaulted key letter isn't changed on
an element rename.

Added Deployment, External Entity, and ModelClass to the defaulted key
letter rename tests.