Eliminate obsolete tests
This commit removes the (now) obsolete formalize and unformalize tests
which are tightly coupled to wizards that no longer exist.  It also
removes tests of the association table editor that never really caught
on and has not been included in the build for a long time.  However,
this commit retains test models and expected results associated with the
formalize and unformalize tests because someday, we may want to add
formalize/unformalize tests that exercise the association editor, and
these models and expected results might then be useful.