Use association editor instead of unformalize CME
The test in question is currently being skipped via @Ignore, so I am
assuming that it is broken in some way.  This commit allows the test to
build successfully in the absence of the unformalize CME removed
by #12599.  Running the test produces a failure when the compare editor
is opened on line 533.  I did not attempt to isolate the failure as part
of this work.